+++ core/src/com/biglybt/core/speedmanager/impl/v2/SpeedManagerAlgorithmProviderV2.java
@@ -154,14 +154,14 @@ SpeedManagerAlgorithmProviderV2
 	}
 
 	@Override
-	public int
+	public long
 	getCurrentChokeSpeed()
 	{
         return strategy.getCurrentChokeSpeed();
 	}
 
 	@Override
-	public int
+	public long
 	getMaxUploadSpeed()
 	{
         return strategy.getMaxUploadSpeed();
